+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 ### enhancements
+  * Check if the given association responds to `order` before calling it. [@laurocaetano](https://github.com/laurocaetano)
   * Add Bootstrap 3 initializer template.
   * For radio or checkbox collection always use `:item_wrapper_tag` to wrap the content and add `label` when using `boolean_style` with `:nested` [@kassio](https://github.com/kassio) and [@erichkist](https://github.com/erichkist)
   * `input_field` uses the same wrapper as input but only with attribute components. [@nashby](https://github.com/nashby)